Some refactoring and work on supporting host certs
