+++ client/cmd/ssh-ca-client-cli/main.go
@@ -1,0 +1,16 @@
+package main
+
+import (
+	"context"
+	"log/slog"
+	"os"
+
+	"github.com/andrewheberle/serverless-ssh-ca/client/internal/pkg/cli"
+)
+
+func main() {
+	if err := cli.Execute(context.Background(), os.Args[1:]); err != nil {
+		slog.Error("error during execution", "error", err)
+		os.Exit(1)
+	}
+}
