--- client/cmd/ssh-ca-client/log.go
@@ -1,13 +1,0 @@
-//go:build !tray
-
-package main
-
-import (
-	"fmt"
-	"os"
-)
-
-func logFatal(format string, a ...any) {
-	fmt.Printf(format, a...)
-	os.Exit(1)
-}
