+++ client/cmd/ssh-ca-client/main_windows.go
@@ -1,13 +1,13 @@
-//go:build tray
-
 //go:generate go-winres make --product-version=git-tag --file-version=git-tag
 
 package main
 
 import (
+	"context"
 	"fmt"
 	"os"
 
+	"github.com/andrewheberle/serverless-ssh-ca/client/internal/pkg/gui"
 	"golang.org/x/sys/windows/svc/eventlog"
 )
 
@@ -23,4 +23,10 @@ func logFatal(format string, a ...any) {
 	}
 
 	os.Exit(1)
+}
+
+func main() {
+	if err := gui.Execute(context.Background(), os.Args[1:]); err != nil {
+		logFatal("Error during execution: %s\n", err)
+	}
 }
