+++ client/internal/pkg/cli/cli.go
@@ -1,11 +1,13 @@
-package cmd
+package cli
 
 import (
+	"context"
 	"errors"
 	"os"
 	"path/filepath"
 
 	"github.com/andrewheberle/serverless-ssh-ca/client/internal/pkg/config"
+	"github.com/andrewheberle/serverless-ssh-ca/client/internal/pkg/config/user"
 	"github.com/andrewheberle/simplecommand"
 	"github.com/bep/simplecobra"
 )
@@ -14,7 +16,7 @@ type rootCommand struct {
 	systemConfigFile string
 	userConfigFile   string
 
-	config *config.Config
+	config config.Config
 
 	*simplecommand.Command
 }
@@ -28,7 +30,7 @@ func (c *rootCommand) Init(cd *simplecobra.Commandeer) error {
 		return err
 	}
 
-	user, system, err := configDirs()
+	user, system, err := config.ConfigDirs()
 	if err != nil {
 		return err
 	}
@@ -51,7 +53,7 @@ func (c *rootCommand) PreRun(this, runner *simplecobra.Commandeer) error {
 	}
 
 	// load config
-	config, err := config.LoadConfig(c.systemConfigFile, c.userConfigFile)
+	config, err := user.LoadConfig(c.systemConfigFile, c.userConfigFile)
 	if err != nil {
 		return err
 	}
@@ -60,4 +62,40 @@ func (c *rootCommand) PreRun(this, runner *simplecobra.Commandeer) error {
 	return nil
 }
 
-// Execute function is in cli.go or tray.go depending on build tags
+func Execute(ctx context.Context, args []string) error {
+	rootCmd := &rootCommand{
+		Command: simplecommand.New("ssh-ca-client-cli", "A CLI based client for a serverless SSH CA"),
+	}
+
+	// set up host subcommand structure
+	hostCmd := simplecommand.New("host", "Handle host keys")
+
+	rootCmd.SubCommands = []simplecobra.Commander{
+		&generateCommand{
+			Command: simplecommand.New("generate", "Generate a SSH private key"),
+		},
+		&loginCommand{
+			Command: simplecommand.New("login", "Login via OIDC and request a certificate from CA"),
+		},
+		&showCommand{
+			Command: simplecommand.New("show", "Show existing private/public key"),
+		},
+		&versionCommand{
+			Command: simplecommand.New("version", "Show the current version of the ssh-ca-client"),
+		},
+		hostCmd,
+	}
+
+	// Set up simplecobra
+	x, err := simplecobra.New(rootCmd)
+	if err != nil {
+		return err
+	}
+
+	// run command with the provided args
+	if _, err := x.Execute(ctx, args); err != nil {
+		return err
+	}
+
+	return nil
+}
