+++ client/internal/pkg/config/host/host_others.go
@@ -1,0 +1,24 @@
+//go:build !windows
+
+package host
+
+import (
+	"fmt"
+	"os"
+
+	"sigs.k8s.io/yaml"
+)
+
+func loadSystemConfig(name string) (SystemConfig, error) {
+	y, err := os.ReadFile(name)
+	if err != nil {
+		return SystemConfig{}, err
+	}
+
+	var config SystemConfig
+	if err := yaml.Unmarshal(y, &config); err != nil {
+		return SystemConfig{}, fmt.Errorf("problem parsing system config: %w", err)
+	}
+
+	return config, nil
+}
