+++ client/internal/pkg/gui/gui_windows.go
@@ -1,6 +1,4 @@
-//go:build tray
-
-package cmd
+package gui
 
 import (
 	"context"
@@ -23,6 +21,17 @@ import (
 
 //go:embed icons
 var resources embed.FS
+
+const appName = "Serverless SSH CA Client"
+
+func configDirs() (user, system string, err error) {
+	dir, err := os.UserConfigDir()
+	if err != nil {
+		return "", "", err
+	}
+
+	return filepath.Join(dir, appName), filepath.Join(os.Getenv("ProgramData"), appName), nil
+}
 
 func runInstall() error {
 	return eventlog.InstallAsEventCreate("Serverless SSH CA Client", eventlog.Error|eventlog.Warning|eventlog.Info)
