+++ client/internal/pkg/gui/icons/error.ico
