+++ client/internal/pkg/gui/icons/error.png
