+++ client/internal/pkg/gui/icons/error128.png
