+++ client/internal/pkg/gui/icons/error256.png
