+++ client/internal/pkg/gui/icons/error32.png
