+++ client/internal/pkg/gui/icons/error64.png
