+++ client/internal/pkg/gui/icons/ok.ico
