+++ client/internal/pkg/gui/icons/ok.png
