+++ client/internal/pkg/gui/icons/warning.ico
