+++ client/internal/pkg/gui/icons/warning.png
