+++ client/internal/pkg/gui/icons/warning128.png
