+++ client/internal/pkg/gui/icons/warning256.png
