+++ client/internal/pkg/gui/icons/warning32.png
