+++ client/internal/pkg/gui/icons/warning64.png
